requirements.txt
asgiref==3.5.2
distlib==0.3.5
Django==4.1
django-ckeditor==6.4.2
django-js-asset==2.0.0
filelock==3.7.1
Pillow==9.2.0
platformdirs==2.5.2
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.16.2
